fp4 model issue check

+++ flux.1-dev_v2.py
@@ -6,7 +6,7 @@ from nunchaku.utils import get_precision
 
 precision = get_precision()  # auto-detect your precision is 'int4' or 'fp4' based on your GPU
 
-torch_dtype = torch.bfloat16  # Auto-selects bfloat16 on Ampere+ GPUs, float16 otherwise
+torch_dtype = torch.float16  # Auto-selects bfloat16 on Ampere+ GPUs, float16 otherwise
 
 transformer = NunchakuFluxTransformer2DModelV2.from_pretrained(
     f"nunchaku-tech/nunchaku-flux.1-dev/svdq-{precision}_r32-flux.1-dev.safetensors", torch_dtype=torch_dtype
